utils: Put durable file replacement in one module
File writing keeps temporary-file creation, atomic replacement, symlink
handling, and directory syncing inside the general file helper.

Several persistence paths need the same sequence for replacing a filesystem
entry without leaving partial data after a crash. Embedding that sequence
in the general file helper makes modules that only need replacement logic
duplicate it or import unrelated file-handling decisions.

This commit begins separating filesystem persistence responsibilities by
extracting regular file replacement, symlink replacement, and directory
syncing into a low-level utility module.

The next commit will test which module provides those operations before
byte buffers start using them for file output.
